print size of data types

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int intType;
 long int longintType;
 long long int longlongintType;
 float floatType;
-printf("Size of a char: %zu byte(s)\n", sizeof(charType));
-printf("Size of an int: %zu byte(s)\n", sizeof(intType));
-printf("Size of a long int %zu byte(s)\n", sizeof(longintType));
-printf("Size of a long long int %zu byte(s)\n", sizeof(longlongintType));
-printf("Size of a float %zu byte(s)\n", sizeof(floatType));
+printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(charType));
+printf("Size of an int: %lu byte(s)\n",(unsigned long)sizeof(intType));
+printf("Size of a long int %lu byte(s)\n",(unsigned long)sizeof(longintType));
+printf("Size of a long long int %lu byte(s)\n",(unsigned long)sizeof(longlongintType));
+printf("Size of a float %lu byte(s)\n",(unsigned long)sizeof(floatType));
 return (0);
 }
